Update and rename iTransfomerFE.py to iTransfomerI.py
I went down a different route. One possible shortcoming I see of in iTrnasformer is that the model doesn't have a concept of what variate is what. All variates are simply sequences of lagged values. They all go through the same embedding and then the transformer is permutation invariant. Thus, the model has no way of understanding what variate is what, if two variates had the same values for the last lookback_len lags, they look identical to iTransformer. I change this by embedding the index of each variate into a 'name_embedding'. Assuming the variates are always entered in the same order, this creates an identifier for the variate. This is then concatenated to the sequence of lagged values before being sent to the embedding for the time series. This might not be ideal, but it was an easy edit and also allows the final embedding to encode interactions between the identifier and the lagged values.